some changes

+++ artifactoryscript.py
@@ -3,69 +3,92 @@ import json
 import argparse
 from configparser import ConfigParser
 
-parser = argparse.ArgumentParser(description="Jenkins server")
+parser = argparse.ArgumentParser(description="JFrog Artifactory")
 parser.add_argument("-a", "--action", metavar="", help="add, delete, list")
 parser.add_argument(
-    "-f", "--file_name", metavar="", help="Enter the file name."
-)
+    "-f", "--file_f", metavar="", help="Enter the file name to open.")
 parser.add_argument(
-    "-r", "--repo", metavar="", help="Enter the repository key in artifactory."
-)
-#parser.add_argument("-r", "--repo_to_list", metavar="", help="Enter repo_key to list the artifacts.")
+    "-n", "--file_name", metavar="", help="Enter the file name for Artifactory.")
+parser.add_argument(
+    "-r", "--repo", metavar="", help="Enter the repository key in Artifactory.")
 args = parser.parse_args()
 
 
 config = ConfigParser()
 config.read("parametersjfrog.ini")
-
 config_data = config["JFROG"]
 host_jfrog = config_data["hostjfrog"]
 username_jfrog = config_data["usernamejfrog"]
 password_jfrog = config_data["passwordjfrog"]
-url_projects = config_data["url_projects"]
 
 
-
-
-
-#authentication-JFROG
+#auth-JFROG
 artifserver = requests.get(host_jfrog, auth = (username_jfrog, password_jfrog))
 print(f'Status(auth): {artifserver.status_code}')
-#print(artifserver.text)
 print('.....................\n')
 
 
+
+
+#function - list repo
 def create_list(repo,host_jfrog,username_jfrog,password_jfrog):
     artiflist = requests.get(f'{host_jfrog}api/storage/{repo}', auth = (username_jfrog, password_jfrog))
-    print(f'Status(list): {artiflist.status_code}\n')
-    datastr = artiflist.text
-    data = json.loads(datastr)
-    datapars = data['children']
-    print(f'Artifacts({repo}):')
-    for x in datapars:
-    	print(x['uri'])
+    statuslist = artiflist.status_code
+    if statuslist == 200:
+        datastr = artiflist.text
+        data = json.loads(datastr)
+        datapars = data['children']
+        print(f'Artifacts({repo}):')
+        for x in datapars:
+            print(x['uri'])
+    else: 
+        print("List-failed.")
 
 
 
+#main part
 action = args.action
 repo = args.repo
-if action == 'add':
+if action and repo and  action == 'add':
     file_name = args.file_name
-    files = {'file': open('artiffile', 'rb')}
-    artifpost = requests.put(f'{host_jfrog}{repo}/{file_name}', auth = (username_jfrog, password_jfrog), files = files)
-    print(f'{host_jfrog}{repo}/{file_name}')
-    print(f'Status(add): {artifpost.status_code}\n')
-    print('.....................\n')
+    file_f = args.file_f
+    if file_name and file_f:
+        files = {'file': open(file_f, 'rb')}
+        artifput = requests.put(f'{host_jfrog}{repo}/{file_name}', auth = (username_jfrog, password_jfrog), files = files)
+        statusput = artifput.status_code
+        if statusput == 201:
+            print(f'The file ({file_f}) is added. {host_jfrog}{repo}/{file_name}')
+            create_list(repo,host_jfrog,username_jfrog,password_jfrog)
+            print('.....................\n')
+        else: 
+            print("Add-failed.")
+    else:
+        print("Enter the file name for Artifactory(-n) and the file name to open(-f).")
 
-elif action == 'list':
+
+
+elif action and repo and action == 'list':
     create_list(repo,host_jfrog,username_jfrog,password_jfrog)
     print('.....................\n')
 
-elif action == 'delete':
+
+elif action and repo and action == 'delete':
     file_name = args.file_name
-    artifdelete = requests.delete(f'{host_jfrog}{repo}/{file_name}', auth = (username_jfrog, password_jfrog))
-    print(f'Status(delete): {artifdelete.status_code}\n')
-    create_list(repo,host_jfrog,username_jfrog,password_jfrog)
-    print('.....................\n')
+    if file_name:
+        artifdelete = requests.delete(f'{host_jfrog}{repo}/{file_name}', auth = (username_jfrog, password_jfrog))
+        statusdelete = artifdelete.status_code
+        if statusdelete == 405 or statusdelete == 204:
+            print(f'The file ({file_name}) is deleted.')
+            create_list(repo,host_jfrog,username_jfrog,password_jfrog)
+            print('.....................\n')
+        else: 
+            print("Delete-failed.")
+    else:
+        print("Enter the file name for Artifactory(-n).")
+
+
+else:
+    print ("Enter the action: (-a) add, delete, list. Enter the repository key in Artifactory (-r). ")
+    print ("Argument '--help' for options.")
